feat: [PAN-5074] update network selector
